impl Segment for other segments
