+++ crates/prune/src/segments/transaction_lookup.rs
@@ -2,19 +2,19 @@ use crate::{
     segments::{PruneInput, PruneOutput, PruneOutputCheckpoint, Segment},
     PrunerError,
 };
+use rayon::prelude::*;
 use reth_db::{database::Database, tables};
-use reth_interfaces::RethResult;
-use reth_primitives::{PruneCheckpoint, PruneSegment};
-use reth_provider::{DatabaseProviderRW, PruneCheckpointWriter, TransactionsProvider};
+use reth_primitives::PruneSegment;
+use reth_provider::{DatabaseProviderRW, TransactionsProvider};
 use tracing::{instrument, trace};
 
 #[derive(Default)]
 #[non_exhaustive]
-pub(crate) struct Receipts;
+pub(crate) struct TransactionLookup;
 
-impl<DB: Database> Segment<DB> for Receipts {
+impl<DB: Database> Segment<DB> for TransactionLookup {
     fn segment(&self) -> PruneSegment {
-        PruneSegment::Receipts
+        PruneSegment::TransactionLookup
     }
 
     #[instrument(level = "trace", target = "pruner", skip(self, provider), ret)]
@@ -23,29 +23,51 @@ impl<DB: Database> Segment<DB> for Receipts {
         provider: &DatabaseProviderRW<'_, DB>,
         input: PruneInput,
     ) -> Result<PruneOutput, PrunerError> {
-        let tx_range = match input.get_next_tx_num_range(provider)? {
+        let (start, end) = match input.get_next_tx_num_range(provider)? {
             Some(range) => range,
             None => {
-                trace!(target: "pruner", "No receipts to prune");
+                trace!(target: "pruner", "No transaction lookup entries to prune");
                 return Ok(PruneOutput::done())
             }
-        };
+        }
+        .into_inner();
+        let tx_range = start..=(end.min(start + input.delete_limit as u64 - 1));
         let tx_range_end = *tx_range.end();
 
-        let mut last_pruned_transaction = tx_range_end;
-        let (pruned, done) = provider.prune_table_with_range::<tables::Receipts>(
-            tx_range,
+        // Retrieve transactions in the range and calculate their hashes in parallel
+        let hashes = provider
+            .transactions_by_tx_range(tx_range.clone())?
+            .into_par_iter()
+            .map(|transaction| transaction.hash())
+            .collect::<Vec<_>>();
+
+        // Number of transactions retrieved from the database should match the tx range count
+        let tx_count = tx_range.count();
+        if hashes.len() != tx_count {
+            return Err(PrunerError::InconsistentData(
+                "Unexpected number of transaction hashes retrieved by transaction number range",
+            ))
+        }
+
+        let mut last_pruned_transaction = None;
+        let (pruned, _) = provider.prune_table_with_iterator::<tables::TxHashNumber>(
+            hashes,
             input.delete_limit,
-            |_| false,
-            |row| last_pruned_transaction = row.0,
+            |row| {
+                last_pruned_transaction = Some(last_pruned_transaction.unwrap_or(row.1).max(row.1))
+            },
         )?;
-        trace!(target: "pruner", %pruned, %done, "Pruned receipts");
+        let done = tx_range_end == end;
+        trace!(target: "pruner", %pruned, %done, "Pruned transaction lookup");
+
+        let last_pruned_transaction = last_pruned_transaction.unwrap_or(tx_range_end);
 
         let last_pruned_block = provider
             .transaction_block(last_pruned_transaction)?
             .ok_or(PrunerError::InconsistentData("Block for transaction is not found"))?
-            // If there's more receipts to prune, set the checkpoint block number to previous,
-            // so we could finish pruning its receipts on the next run.
+            // If there's more transaction lookup entries to prune, set the checkpoint block number
+            // to previous, so we could finish pruning its transaction lookup entries on the next
+            // run.
             .checked_sub(if done { 0 } else { 1 });
 
         Ok(PruneOutput {
@@ -57,35 +79,18 @@ impl<DB: Database> Segment<DB> for Receipts {
             }),
         })
     }
-
-    fn save_checkpoint(
-        &self,
-        provider: &DatabaseProviderRW<'_, DB>,
-        checkpoint: PruneCheckpoint,
-    ) -> RethResult<()> {
-        provider.save_prune_checkpoint(PruneSegment::Receipts, checkpoint)?;
-
-        // `PruneSegment::Receipts` overrides `PruneSegment::ContractLogs`, so we can preemptively
-        // limit their pruning start point.
-        provider.save_prune_checkpoint(PruneSegment::ContractLogs, checkpoint)?;
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::segments::{PruneInput, PruneOutput, Receipts, Segment};
+    use crate::segments::{PruneInput, PruneOutput, Segment, TransactionLookup};
     use assert_matches::assert_matches;
     use itertools::{
         FoldWhile::{Continue, Done},
         Itertools,
     };
     use reth_db::tables;
-    use reth_interfaces::test_utils::{
-        generators,
-        generators::{random_block_range, random_receipt},
-    };
+    use reth_interfaces::test_utils::{generators, generators::random_block_range};
     use reth_primitives::{BlockNumber, PruneCheckpoint, PruneMode, PruneSegment, TxNumber, B256};
     use reth_provider::PruneCheckpointReader;
     use reth_stages::test_utils::TestTransaction;
@@ -99,14 +104,13 @@ mod tests {
         let blocks = random_block_range(&mut rng, 1..=10, B256::ZERO, 2..3);
         tx.insert_blocks(blocks.iter(), None).expect("insert blocks");
 
-        let mut receipts = Vec::new();
+        let mut tx_hash_numbers = Vec::new();
         for block in &blocks {
             for transaction in &block.body {
-                receipts
-                    .push((receipts.len() as u64, random_receipt(&mut rng, transaction, Some(0))));
+                tx_hash_numbers.push((transaction.hash, tx_hash_numbers.len() as u64));
             }
         }
-        tx.insert_receipts(receipts.clone()).expect("insert receipts");
+        tx.insert_tx_hash_numbers(tx_hash_numbers.clone()).expect("insert tx hash numbers");
 
         assert_eq!(
             tx.table::<tables::Transactions>().unwrap().len(),
@@ -114,7 +118,7 @@ mod tests {
         );
         assert_eq!(
             tx.table::<tables::Transactions>().unwrap().len(),
-            tx.table::<tables::Receipts>().unwrap().len()
+            tx.table::<tables::TxHashNumber>().unwrap().len()
         );
 
         let test_prune = |to_block: BlockNumber, expected_result: (bool, usize)| {
@@ -122,16 +126,16 @@ mod tests {
             let input = PruneInput {
                 previous_checkpoint: tx
                     .inner()
-                    .get_prune_checkpoint(PruneSegment::Receipts)
+                    .get_prune_checkpoint(PruneSegment::TransactionLookup)
                     .unwrap(),
                 to_block,
                 delete_limit: 10,
             };
-            let segment = Receipts::default();
+            let segment = TransactionLookup::default();
 
             let next_tx_number_to_prune = tx
                 .inner()
-                .get_prune_checkpoint(PruneSegment::Receipts)
+                .get_prune_checkpoint(PruneSegment::TransactionLookup)
                 .unwrap()
                 .and_then(|checkpoint| checkpoint.tx_number)
                 .map(|tx_number| tx_number + 1)
@@ -178,11 +182,11 @@ mod tests {
                 last_pruned_block_number.checked_sub(if result.done { 0 } else { 1 });
 
             assert_eq!(
-                tx.table::<tables::Receipts>().unwrap().len(),
-                receipts.len() - (last_pruned_tx_number + 1)
+                tx.table::<tables::TxHashNumber>().unwrap().len(),
+                tx_hash_numbers.len() - (last_pruned_tx_number + 1)
             );
             assert_eq!(
-                tx.inner().get_prune_checkpoint(PruneSegment::Receipts).unwrap(),
+                tx.inner().get_prune_checkpoint(PruneSegment::TransactionLookup).unwrap(),
                 Some(PruneCheckpoint {
                     block_number: last_pruned_block_number,
                     tx_number: Some(last_pruned_tx_number as TxNumber),
